requirements.txt
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
bidict==0.22.0
black==22.6.0
click==8.1.3
Deprecated==1.2.13
Django==4.1
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pyparsing==3.0.9
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
sniffio==1.2.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
wrapt==1.14.1
